requirements.txt
python-dotenv>=1.0.0
pydantic>=2.0.0
pypdf>=4.0.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
networkx>=3.2.0
numpy>=1.26.0
pandas>=2.2.0
requests>=2.31.0
streamlit>=1.32.0
pytest>=8.0.0
